My first rather good setup for guiding for my mesu-200 with DC servos.
Feed-forward is crucial for guiding (through AXISn_SERVO_VELOCITY_FACTOR). Current pwm is at around 4Khz with 15bit resolution.

The DEC PID has not been optimized. Also it seems that for my setup the motor are more loaded in the DEC rather than the RA.

Commiting this for future reference